Refactor session.c into multiple files
