+++ include/aws/cryptosdk/private/session.h
@@ -20,6 +20,9 @@
 #include <aws/cryptosdk/private/header.h>
 #include <aws/cryptosdk/private/cipher.h>
 
+#define DEFAULT_FRAME_SIZE (256 * 1024)
+#define MAX_FRAME_SIZE 0xFFFFFFFF
+
 enum session_state {
 /*** Common states ***/
 
@@ -75,6 +78,36 @@ struct aws_cryptosdk_session {
     /* Decrypted, derived (if applicable) content key */
     struct content_key content_key;
 };
+
+
+/* Common session routines */
+
+void session_change_state(struct aws_cryptosdk_session *session, enum session_state new_state);
+int fail_session(struct aws_cryptosdk_session *session, int error_code);
+
+/* Decrypt path */
+int unwrap_keys(struct aws_cryptosdk_session * restrict session);
+int try_parse_header(
+    struct aws_cryptosdk_session * restrict session,
+    struct aws_byte_cursor * restrict input
+);
+int try_decrypt_body(
+    struct aws_cryptosdk_session * restrict session,
+    struct aws_byte_cursor * restrict poutput,
+    struct aws_byte_cursor * restrict pinput
+);
+
+/* Encrypt path */
+void encrypt_compute_body_estimate(struct aws_cryptosdk_session *session);
+
+int try_gen_key(struct aws_cryptosdk_session *session);
+int try_write_header(struct aws_cryptosdk_session *session, struct aws_byte_cursor *output);
+int try_encrypt_body(
+    struct aws_cryptosdk_session * restrict session,
+    struct aws_byte_cursor * restrict poutput,
+    struct aws_byte_cursor * restrict pinput
+);
+
 
 
 #endif
